mt integration tests - auth

+++ multitenancy/tenant-manager-client/src/main/java/io/apicurio/multitenant/client/Auth.java
@@ -54,6 +54,6 @@ public class Auth {
     }
 
     private boolean isTokenExpired() {
-        return (accessTokenParsed.getExp() != null && accessTokenParsed.getExp() != 0L) && (long) Time.currentTime() > accessTokenParsed.getExp();
+        return (accessTokenParsed.getExp() != null && accessTokenParsed.getExp() != 0L) && Time.currentTime() > accessTokenParsed.getExp();
     }
 }
